Add an out for recursively defined protobuf
There is no limitation on protobufs being described in a recursive fashion (A has a field B which has a field A)
I added a seenMap, which records down the stack how many occurances there have been of a field name. If we encounter the same field
a greater number of time than our max recursion, we don't add any more of it's subfields, and we end the struct there.

I also added the ability to ignore the error field that's added

Finally, I have certain needs with a protobuf object with any Any field, where that Any field will be another protobuf definition.
The schema is generic, so it doesn't know what the any field is, but at runtime I know, so I want to manually set what the schema should be

For that reason, I add an innerDesc, which is the descriptor that is contained in the Any field, and a connectingField, which is the name of the Any field that will be merged.